refactor(cli): split cli.rs integration tests by subcommand
Split the 1745-line monolithic test file into focused modules:
- common/mod.rs: shared helpers (rsigma(), temp_file()) and fixtures
- cli_parse.rs: parse, condition, stdin subcommands
- cli_validate.rs: validate subcommand
- cli_eval.rs: eval subcommand (detection, pipeline, jq, correlation, filter)
- cli_lint.rs: lint subcommand including --fix tests
- cli_daemon.rs: daemon state persistence and streaming pipeline tests

Single-use fixture constants are colocated with their test file to
avoid dead_code warnings across integration test binaries.
